rename users service to auth service
